refactor collect_bugs
