+++ gitbugactions/collect_bugs/test_config.py
@@ -1,0 +1,2 @@
+class TestConfig:
+    normalize_non_code_patch = True
